update: コンポーネントの更新
